change in table for loans and extra_repayment_schedules
- add payment after extra column for loans table
- remove unnecessary columns in extra_repayment_schedules table
- fix factory and model with changes

+++ database/migrations/2024_01_04_083121_create_loan_amortization_schedules_table.php
@@ -12,7 +12,7 @@ return new class extends Migration {
     {
         Schema::create('loan_amortization_schedules', function (Blueprint $table) {
             $table->id();
-            $table->tinyInteger('month_number');
+            $table->integer('month_number');
             $table->double('starting_balance');
             $table->double('monthly_payment');
             $table->double('principal_component');
